[feat] Refactor tools base.py (#175)
* changes

* fix test

* fix test

* add temp readme

* styles
