+++ erniebot-agent/tests/unit_tests/tools/test_file_in_tool.py
@@ -31,7 +31,7 @@ from requests.models import Response
 
 from erniebot_agent.file_io.file_manager import FileManager
 from erniebot_agent.tools import RemoteToolkit
-from erniebot_agent.tools.base import (
+from erniebot_agent.tools.utils import (
     get_file_info_from_param_view,
     parse_file_from_json_response,
 )
