+++ erniebot-agent/tests/unit_tests/tools/test_schema.py
@@ -23,13 +23,13 @@ from openapi_spec_validator.readers import read_from_filename
 from pydantic import Field
 
 from erniebot_agent.tools import RemoteToolkit
-from erniebot_agent.tools.base import tool_response_contains_file
 from erniebot_agent.tools.schema import (
     ToolParameterView,
     get_typing_list_type,
     is_optional_type,
     json_type,
 )
+from erniebot_agent.tools.utils import tool_response_contains_file
 from erniebot_agent.utils.common import create_enum_class
